Title: 체질량지수, Time: 142ms, Memory: 10MB, Status: Wrong Answer - CodeTree

+++ 240209/체질량지수/bmi.java
@@ -9,12 +9,12 @@ public class Main {
 
 		    int bmi = w*10000 / (h * h);
 
-        if (bmi >= 25) {
-            System.out.println((int)bmi);
-            System.out.println("Obesity");
-        }
-        else if (bmi < 25) {
-            System.out.println("pass");
-        }
+            if (bmi >= 25) {
+                System.out.println(bmi);
+                System.out.println("Obesity");
+            }
+            else if (bmi < 25) {
+                System.out.println("pass");
+            }
     }
 }
